- move row sorting/merging utils away from actions #2014 - fetch quick actions after websocket event #2014
